fix(notebooks): move funnel series and exclusion selectors
