+++ frontend/src/queries/nodes/InsightViz/InsightContainer.tsx
@@ -61,9 +61,8 @@ export function InsightContainer({
 
     const { activeView } = useValues(insightNavLogic(insightProps))
 
-    const { isFunnelWithEnoughSteps, hasFunnelResults, areExclusionFiltersValid } = useValues(
-        funnelDataLogic(insightProps)
-    )
+    const { hasFunnelResults } = useValues(funnelDataLogic(insightProps))
+    const { isFunnelWithEnoughSteps, areExclusionFiltersValid } = useValues(insightVizDataLogic(insightProps))
     const {
         isTrends,
         isFunnels,
